Create JavaFX project in IntelliJ and create config for IDE with documentation
