+++ .github/ISSUE_TEMPLATE/feature.md
@@ -1,14 +1,15 @@
 ---
 name: Feature
 about: A main feature of the project
-title: "[FEATURE] "
+title: '[FEATURE] '
 labels: feature
 assignees: ''
-
 ---
 
 # Description
-- [ e.g. The UI needs to display...]
 
-# Finished when...
-- [ e.g. There is UI for the user to interact with the project ]
+-   [ e.g. The UI needs to display...]
+
+# Finished Criteria
+
+-   [ e.g. There is UI for the user to interact with the project ]
