+++ .github/ISSUE_TEMPLATE/task.md
@@ -1,14 +1,19 @@
 ---
 name: Task
 about: A work to be done related to the feature
-title: "[TASK] "
+title: '[TASK] '
 labels: task
 assignees: ''
-
 ---
 
 # Description
-- [ e.g. The user needs to see data from the database]
 
-# Finished when...
-- [ e.g. Button is clicked, UI can display data from the database ]
+-   [ e.g. The user needs to see data from the database]
+
+**Tech Notes:**
+
+-   [ Any notes that might help the developer tackle this task ]
+
+# Finished Criteria
+
+-   [ e.g. Button is clicked, UI can display data from the database ]
