+++ .idea/compiler.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="CompilerConfiguration">
+    <bytecodeTargetLevel>
+      <module name="ui" target="25" />
+      <module name="ui.main" target="25" />
+      <module name="ui.test" target="25" />
+    </bytecodeTargetLevel>
+  </component>
+</project>
