+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+    "gradle.nestedProjects": true,
+    "java.format.settings.url": ".vscode/java-formatter.xml",
+    "[java]": {
+        "editor.tabSize": 4
+    },
+    "java.configuration.updateBuildConfiguration": "automatic",
+    "editor.formatOnSave": true
+}
