+++ application/src/main/java/org/utdteamthreefive/backend/SampleClass.java
@@ -1,0 +1,7 @@
+package org.utdteamthreefive.backend;
+
+public class SampleClass {
+    public String getGreeting() {
+        return "Welcome to JavaFX Application!";
+    }
+}
