--- ui/src/main/java/com/utd/team35/ui/HelloController.java
@@ -1,14 +1,0 @@
-package com.utd.team35.ui;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class HelloController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-}
