--- ui/src/main/java/module-info.java
@@ -1,8 +1,0 @@
-module com.utd.team35.ui {
-    requires javafx.controls;
-    requires javafx.fxml;
-
-
-    opens com.utd.team35.ui to javafx.fxml;
-    exports com.utd.team35.ui;
-}
